+++ jam.mjs
@@ -1,5 +1,5 @@
 #!/usr/bin/env node
-import record from './src/record.mjs';
+import entrypoint from './src/entrypoint.mjs';
 import * as constants from './src/constants.mjs';
 import { enable as enableDiagnostics } from './src/diagnostics.js';
 
@@ -14,7 +14,7 @@ const run = async (argv, cwd) => {
     process.exit(0);
   }
 
-  record({
+  entrypoint({
     script,
     cwd,
     log: console.log,
@@ -24,6 +24,10 @@ const run = async (argv, cwd) => {
     .then((result) => {
       if (result.browser) {
         console.log('Browser launched');
+        // TODO: Figure out why spawning the browser process leaves
+        // our original node process running.
+        // Maybe we should always exit even for process launches
+        process.exit(0);
       }
 
       if (result.process) {